requirements.txt
django
pymysql
djangorestframework
